backend/requirements.txt
# FastAPI и сервер
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Базы данных
redis>=5.0.1
psycopg2-binary>=2.9.9
pgvector>=0.2.5
sqlalchemy>=2.0.23
alembic>=1.12.1  # Для миграций

# HTTP клиенты
aiohttp>=3.9.1
requests>=2.31.0

# Научные вычисления
numpy>=1.24.3
pandas>=2.1.3

# Парсинг
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Конфигурация
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Дата и время
python-dateutil>=2.8.2

# Для работы с эмбеддингами (альтернатива, если Gigachat не работает)
sentence-transformers>=2.2.2
torch>=2.1.0
transformers>=4.35.0

# Логирование
loguru>=0.7.2

# Для UUID
uuid>=1.30